require sudo for build, remove docker image

+++ create-ccu-image.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
+docker rm -v pigen_work
+
 pushd stage2-custom/04-ccu-app/files
 rm -rf DCU-central-control-app
 git clone --single-branch --branch beta https://github.com/MrAlfabet/DCU-central-control-app.git
